+++ src/edu/antonym/test/TestCase2.java
@@ -50,7 +50,7 @@ public class TestCase2 implements EmbeddingEvaluator {
 			
 			while(line!=null){
 				String[] words = line.split("\\s+");
-				ArrayList<float[]> vectors = new ArrayList<float[]>();
+				ArrayList<double[]> vectors = new ArrayList<double[]>();
 				for(int j=0;j<words.length;j++){
 					int wordID = vocab.lookupWord(words[j]);
 					if (wordID != OOV) {
